Refactor notification state
Since the framework handles conversation style notifications that have independent lines with author/message information, we want to model our system after it to make things convenient.
Instead of assembling message information ourselves, we can just have a collection of active conversations in ConversationsInfo, have each conversation represented by a Conversation, and have each message be represented by MessageLineInfo. The general architecture was there with how wearable notifications existed, but this just unifies the functionality.
Notably, we make sure to associate attachment uris with each message instead of having one attachment uri per conversation and the same goes for the author of each message where we can have a new avatar for each line.

Additionally, I added extra text that can appear in a notification based on the file type received or if we don't know what to do with it.