Refactor notification updates
This continues to remove the old custom notification functionality and moves into the model mentioned in the MessageNotificationState refactor. Now instead of updating all notifications and selectively buzzing the user's device, we only update notifications for conversations that have updates as they come in. With the changes made to MessageNotificationState, sending notifications becomes very straightforward.

Additionally, since we get to merge the wearable code into mobile, we get inline replies